F add auth (#14)
* added port for mongo (#11)


* Update README.md

* Update README.md

* merge

* add tiptap

* fix

+++ src/resolvers/course.resolver.ts
@@ -13,9 +13,12 @@ import {
   ObjectType,
   FieldResolver,
   Root,
+  UseMiddleware,
+  Ctx,
 } from 'type-graphql'
 import { Converter } from 'showdown'
-import { CourseInput, CourseUpdateInput } from '../Entity/course/course.input'
+import { IContext, isAuth } from '../middlewares/auth.middleware'
+import CourseInput, { CourseUpdateInput } from '../Entity/course/course.input'
 import { CourseModel, Course } from '../Entity/course/course.entity'
 import { ClassRoom, ClassRoomModel } from '../Entity/classes/class.entity'
 
@@ -61,7 +64,17 @@ class CourseResolver {
   }
 
   @Mutation(() => Course)
-  async createCourse(@Arg('data') data: CourseInput): Promise<Course> {
+  @UseMiddleware(isAuth)
+  async createCourse(
+    @Ctx() ctx: IContext,
+    @Arg('data') data: CourseInput
+  ): Promise<Course> {
+    const { user } = ctx.payload!
+    const classRoom = await ClassRoomModel.findById(data.classRoom)
+    if (user._id.toString() !== classRoom?.owner?.toString()) {
+      throw new Error('Promision denied')
+    }
+
     const converter = new Converter()
     data.steps.forEach((step) => {
       step.contentHtml = converter.makeHtml(step.contentMd)
@@ -88,11 +101,16 @@ class CourseResolver {
     return course
   }
 
-  @Mutation(() => IdeleteResponse)
-  async deleteCourse(@Arg('id') id: string): Promise<IdeleteResponse> {
-    const course = CourseModel.findById(id)
-    const deletedCourse = await course.deleteOne()
-    return deletedCourse
+  @Mutation(() => Boolean)
+  async deleteCourse(@Arg('id') id: string): Promise<boolean> {
+    const course = await CourseModel.findById(id)
+    console.log('course', course)
+    if (course) {
+      const deletedCourse = await course.deleteOne()
+      console.log(deletedCourse)
+      return true
+    }
+    return false
   }
 }
 export default CourseResolver
